Migrate Manage Reusable Block to Playwright
