+++ test/e2e/assets/greeting-reusable-block.json
